Replace Voice hashmap with VecDeque, limit voice count. (#26)
Use a `VecDeque`, as it will internally use a circular queue and avoid allocations.

There's work to be done here on dealing with new incoming voices. For now, we don't allow users to add new voices (unless they are already playing that note, in which case we release the old voice and replace it with a new one).

We'd like to eventually add mono and legato modes, as well as portamento.